+++ src/main/java/fr/estecka/variantscit/commands/ModuleCommands.java
@@ -6,6 +6,8 @@ import com.mojang.brigadier.exceptions.CommandSyntaxException;
 import com.mojang.brigadier.suggestion.Suggestions;
 import com.mojang.brigadier.suggestion.SuggestionsBuilder;
 import fr.estecka.variantscit.VariantsCitMod;
+import fr.estecka.variantscit.assetgen.GeneratedResourcePack;
+import fr.estecka.variantscit.assetgen.TemplatedResource;
 import fr.estecka.variantscit.modules.IBakedModule;
 import fr.estecka.variantscit.reload.EModuleContext;
 import fr.estecka.variantscit.reload.MetaModule;
@@ -36,6 +38,7 @@ public class ModuleCommands
 {
 	static public final Identifier ID = Identifier.of(VariantsCitMod.MODID, "modules");
 
+	static public final String ASSET_ARG   = "asset id";
 	static public final String CONTEXT_ARG = "context";
 	static public final String MODULE_ARG  = "module id";
 	
@@ -56,8 +59,15 @@ public class ModuleCommands
 			.then(module)
 			;
 
+		var assetgen = literal("assetgen")
+			.then(argument(ASSET_ARG, identifier())
+				.suggests(ModuleCommands::AssetAutofill)
+				.executes(ModuleCommands::AssetDump)
+			);
+
 		var root = literal(VariantsCitMod.MODID)
 			.then(context)
+			.then(assetgen)
 			;
 
 		dispatcher.register(root);
@@ -67,6 +77,11 @@ public class ModuleCommands
 /******************************************************************************/
 /* # Autofill                                                                 */
 /******************************************************************************/
+
+	static private CompletableFuture<Suggestions> AssetAutofill(final CommandContext<FabricClientCommandSource> context, final SuggestionsBuilder builder){
+		CommandSource.suggestIdentifiers(GeneratedResourcePack.INSTANCE.Get().keySet(), builder);
+		return builder.buildFuture();
+	}
 
 	static private CompletableFuture<Suggestions> ContextAutofill(final CommandContext<FabricClientCommandSource> context, final SuggestionsBuilder builder){
 		for (EModuleContext moduleContext : EModuleContext.values())
@@ -117,6 +132,16 @@ public class ModuleCommands
 
 		CommandLogger logger = new CommandLogger(context, moduleContext, meta);
 		return command.Execute(context, logger, module);
+	}
+
+	static private int AssetDump(CommandContext<FabricClientCommandSource> context) throws CommandSyntaxException {
+		Identifier id = context.getArgument(ASSET_ARG, Identifier.class);
+		TemplatedResource resource = (TemplatedResource)GeneratedResourcePack.INSTANCE.Get().get(id);
+
+		VariantsCitMod.LOGGER.info("{}:\n{}", id, resource.getString());
+
+		context.getSource().sendFeedback(Text.literal("Asset content was printed into the game's log."));
+		return 0;
 	}
 
 	static private int Dump(CommandContext<FabricClientCommandSource> context, CommandLogger logger, IBakedModule module){